LayoutTests/imported/w3c/web-platform-tests/cssom/serialize-variable-reference-expected.txt


PASS Longhand with variable preserves original serialization: with withespace 
PASS Shorthand with variable preserves original serialization: with withespace 
FAIL Longhand with variable preserves original serialization: without withespace assert_equals: expected "font-size:var(--a);" but got "font-size: var(--a);"
FAIL Shorthand with variable preserves original serialization: without withespace assert_equals: expected "font:var(--a);" but got "font: var(--a);"
